Widen the store mask chain to 64 bits and extend the fa18c carriage grammar (#27): dual AMRAAM rails and inboard heaters. The catalog grows past 32 entries with the twin racks on stations 3/7, the AMRAAM pair rounds, and the inboard AIM-9 singles and pairs, so Airframe.Default, Model.Stores, stores_mask, attach, armed and the wasm boundary all move from uint32 to uint64 - a 32-bit mask silently dropped the new bits server-side and would have aliased them client-side, and tests now pin entries past bit 31 as tripwires. Every rail point is a LAU-127 carrying either family; the stores entries collapse to generic store-plus-station naming with twin points suffixed a/b outer-first; the 9M SMS order gains its inboard ring after the outboard one, starboard seeding. Mixed twin pairs are real LAU-115 carriage and survive the grant; the centreline still never carries a missile.

+++ games/air/air.go
@@ -793,7 +793,7 @@ func (i *instance) Step(tick uint64, inputs map[int][]game.Input) {
 // order and the other rounds stay carried (bot.go calls this on every
 // launch; the discipline itself is #25's re-sweep). Loadout-less crafts'
 // attach() shares it as the legacy fallback.
-func armed(count int) uint32 {
+func armed(count int) uint64 {
 	fired := shots - count
 	if fired < 0 {
 		fired = 0
